First work on the deletion of the metrics
