+++ .github/workflows/ci.yaml
@@ -64,6 +64,7 @@ jobs:
         tox-environments:
           - integration-charm
           - integration-provider
+          - integration-observability
     name: ${{ matrix.tox-environments }}
     needs:
       - lint
