+++ tests/integration/helpers.py
@@ -4,10 +4,16 @@
 
 import json
 import logging
-from typing import Dict
+import subprocess
+from time import sleep
+from typing import Any, Dict, MutableMapping
 
+import boto3
+from botocore.client import Config
 from juju.unit import Unit
 from pytest_operator.plugin import OpsTest
+
+BUCKET_NAME = "test-bucket"
 
 logger = logging.getLogger(__name__)
 
@@ -61,3 +67,127 @@ async def update_juju_secret(
         logger.warning(
             f"Update Juju secret exited with non zero status. \nSTDOUT: {stdout.strip()} \nSTDERR: {stderr.strip()}"
         )
+
+
+def run_service_account_registry(*args):
+    """Run service_account_registry CLI command with given set of args.
+
+    Returns:
+        Tuple: A tuple with the content of stdout, stderr and the return code
+            obtained when the command is run.
+    """
+    command = ["python3", "-m", "spark8t.cli.service_account_registry", *args]
+    try:
+        output = subprocess.run(command, check=True, capture_output=True)
+        return output.stdout.decode(), output.stderr.decode(), output.returncode
+    except subprocess.CalledProcessError as e:
+        return e.stdout.decode(), e.stderr.decode(), e.returncode
+
+
+def get_secret_data(namespace: str, secret_name: str):
+    """Retrieve secret data for a given namespace and secret."""
+    command = ["kubectl", "get", "secret", "-n", namespace, "--output", "json"]
+    try:
+        output = subprocess.run(command, check=True, capture_output=True)
+        # output.stdout.decode(), output.stderr.decode(), output.returncode
+        result = output.stdout.decode()
+        logger.info(f"Command: {command}")
+        logger.info(f"Secrets for namespace: {namespace}")
+        logger.info(f"Request secret: {secret_name}")
+        logger.info(f"results: {str(result)}")
+        secrets = json.loads(result)
+        data = {}
+        for secret in secrets["items"]:
+            name = secret["metadata"]["name"]
+            logger.info(f"\t secretName: {name}")
+            if name == secret_name:
+                data = {}
+                if "data" in secret:
+                    data = secret["data"]
+        return data
+    except subprocess.CalledProcessError as e:
+        return e.stdout.decode(), e.stderr.decode(), e.returncode
+
+
+async def juju_sleep(ops: OpsTest, app_name: str, time: int):
+    await ops.model.wait_for_idle(
+        apps=[
+            app_name,
+        ],
+        idle_period=time,
+        timeout=300,
+    )
+
+
+def setup_s3_bucket_for_sch_server(endpoint_url: str, aws_access_key: str, aws_secret_key: str):
+    config = Config(connect_timeout=60, retries={"max_attempts": 0})
+    session = boto3.session.Session(
+        aws_access_key_id=aws_access_key, aws_secret_access_key=aws_secret_key
+    )
+    s3 = session.client("s3", endpoint_url=endpoint_url, config=config)
+    # delete test bucket and its content if it already exist
+    buckets = s3.list_buckets()
+    for bucket in buckets["Buckets"]:
+        bucket_name = bucket["Name"]
+        if bucket_name == BUCKET_NAME:
+            logger.info(f"Deleting bucket: {bucket_name}")
+            objects = s3.list_objects_v2(Bucket=BUCKET_NAME)["Contents"]
+            objs = [{"Key": x["Key"]} for x in objects]
+            s3.delete_objects(Bucket=BUCKET_NAME, Delete={"Objects": objs})
+            s3.delete_bucket(Bucket=BUCKET_NAME)
+
+    logger.info("create bucket in minio")
+    for i in range(0, 30):
+        try:
+            s3.create_bucket(Bucket=BUCKET_NAME)
+            break
+        except Exception as e:
+            if i >= 30:
+                logger.error(f"create bucket failed....exiting....\n{str(e)}")
+                raise
+            else:
+                logger.warning(f"create bucket failed....retrying in 10 secs.....\n{str(e)}")
+                sleep(10)
+                continue
+
+    s3.put_object(Bucket=BUCKET_NAME, Key=("spark-events/"))
+    logger.debug(s3.list_buckets())
+
+
+async def run_action(
+    ops_test: OpsTest, action_name: str, params: Dict[str, str], app_name: str, num_unit=0
+) -> Any:
+    """Use the charm action to start a password rotation."""
+    action = await ops_test.model.units.get(f"{app_name}/{num_unit}").run_action(
+        action_name, **params
+    )
+    password = await action.wait()
+    return password.results
+
+
+def flatten(map: MutableMapping, parent: str = "", separator: str = ".") -> dict[str, str]:
+    """Flatten given nested dictionary to a non-nested dictionary where keys are separated by a dot.
+
+    For example, consider a nested dictionary as follows:
+
+        {
+            'foo': {
+                'bar': 'val1',
+                'grok': 'val2'
+            }
+        }
+
+    The return value would be as follows:
+        {
+            'foo.bar': 'val1',
+            'foo.grok': 'val2'
+        }
+    """
+    items = []
+    for key, value in map.items():
+        new_key = parent + separator + key if parent else key
+        if isinstance(value, MutableMapping):
+            items.extend(flatten(value, new_key, separator).items())
+        else:
+            items.append((new_key, value))
+    return dict(items)
