+++ tox.ini
@@ -23,6 +23,7 @@ set_env =
     PY_COLORS=1
     charm: TEST_FILE=test_charm.py
     provider: TEST_FILE=test_provider.py
+    observability: TEST_FILE=test_observability.py
 pass_env =
     PYTHONPATH
     CHARM_BUILD_DIR
@@ -72,7 +73,7 @@ commands =
         -m pytest -vv --tb native -s {posargs} {[vars]tests_path}/unit
     poetry run coverage report
 
-[testenv:integration-{charm,provider}]
+[testenv:integration-{charm,provider,observability}]
 description = Run integration tests
 pass_env =
     {[testenv]pass_env}
